Move `TaskId` to core
