+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/TaskIdDeserializer.java
@@ -19,6 +19,7 @@
 package ai.grakn.engine.tasks;
 
 
+import ai.grakn.engine.TaskId;
 import org.apache.kafka.common.serialization.Deserializer;
 
 import java.nio.charset.StandardCharsets;
