+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/TaskState.java
@@ -18,6 +18,7 @@
 
 package ai.grakn.engine.tasks;
 
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.TaskStatus;
 import mjson.Json;
