+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/config/ConfigHelper.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.engine.tasks.config;
 
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.util.ConfigProperties;
 import org.apache.curator.framework.CuratorFramework;
