+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/manager/StandaloneTaskManager.java
@@ -20,7 +20,7 @@
 package ai.grakn.engine.tasks.manager;
 
 import ai.grakn.engine.tasks.BackgroundTask;
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskManager;
 import ai.grakn.engine.tasks.TaskSchedule;
 import ai.grakn.engine.tasks.TaskState;
