+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/storage/TaskStateZookeeperStore.java
@@ -19,7 +19,7 @@
 package ai.grakn.engine.tasks.storage;
 
 import ai.grakn.engine.TaskStatus;
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.tasks.TaskStateStorage;
 import ai.grakn.engine.tasks.manager.ZookeeperConnection;
