+++ grakn-test/src/test/java/ai/grakn/generator/TaskStates.java
@@ -21,7 +21,7 @@ package ai.grakn.generator;
 
 import ai.grakn.engine.TaskStatus;
 import ai.grakn.engine.tasks.BackgroundTask;
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskSchedule;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.test.engine.tasks.FailingTestTask;
