+++ grakn-test/src/test/java/ai/grakn/test/engine/tasks/manager/multiqueue/MultiQueueTaskRunnerTest.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.test.engine.tasks.manager.multiqueue;
 
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskSchedule;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.tasks.TaskStateStorage;
