+++ grakn-test/src/test/java/ai/grakn/test/engine/tasks/manager/singlequeue/SingleQueueTaskRunnerTest.java
@@ -19,7 +19,7 @@
 
 package ai.grakn.test.engine.tasks.manager.singlequeue;
 
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.tasks.manager.singlequeue.SingleQueueTaskRunner;
 import ai.grakn.engine.tasks.storage.TaskStateInMemoryStore;
