+++ grakn-test/src/test/java/ai/grakn/test/engine/tasks/storage/TaskStateInMemoryStoreTest.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.test.engine.tasks.storage;
 
-import ai.grakn.engine.tasks.TaskId;
+import ai.grakn.engine.TaskId;
 import ai.grakn.engine.tasks.TaskSchedule;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.tasks.TaskStateStorage;
